Start moving format-specific code into doctest submodule
