+++ src/librustdoc/doctest/markdown.rs
@@ -1,0 +1,1 @@
+//! Doctest functionality used only for doctests in `.md` Markdown files.
